api/requirements.txt
boto3==1.28.38
psycopg2-binary==2.9.7
uvicorn==0.23.2
fastapi==0.103.0
loguru==0.7.0
python-decouple==3.8
python-multipart==0.0.6
Pillow==10.0.0
pyjwt[crypto]==2.8.0
SQLAlchemy==2.0.23